+++ app/src/main/java/es/soutullo/blitter/view/activity/ManualTranscriptionActivity.kt
@@ -63,7 +63,7 @@ class ManualTranscriptionActivity : AppCompatActivity() {
         val productPriceText = this.findViewById<EditText>(R.id.product_price_field)
 
         if (productNameText.text.isNotBlank() && productPriceText.text.isNotBlank() && productPriceText.text.toString().toFloatOrNull() != null) {
-            this.productsAdapter.add(ManualTranscriptionProduct(productNameText.text.trim().toString(), productPriceText.text.toString().toFloat(), 1))
+            this.productsAdapter.add(ManualTranscriptionProduct(productNameText.text.trim().toString(), productPriceText.text.toString().toDouble(), 1))
             this.productsAdapter.recyclerView?.scrollToPosition(this.productsAdapter.itemCount - 1)
             this.doBackup()
